move menu code
